neatlogs/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-semantic-conventions
openinference-instrumentation